Rename AWS client file.
