+++ graphql/integration-tests/utils/aws-client-dynamodb.js
